pyproject.toml
[tool.poetry]
name = "seed-qe-automation-framework"
version = "0.1.0"
description = "MVP build for SEED FY25_Q3 data verification tool"
authors = ["pthi1650 <156160259+pthi1650@users.noreply.github.com>"]
package-mode = false  # Add this line to disable package mode
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
polars = { version = "^0.20.31", extras = ["pyarrow"] }
pytest = "^8.2.2"
toml = "^0.10.2"
pyyaml = "^6.0.1"
boto3 = "^1.34.128"
sqlalchemy = "1.4.48"
pytest-xdist = "^3.6.1"
psycopg2 = "^2.9.9"
redshift-connector = "2.1.2"
sqlalchemy-redshift = "^0.8.14"
pyarrow = "^16.1.0"
pandas = "^2.2.2"
black = "^24.4.2"
aiofiles = "24.1.0"
paramiko = "^3.4.1"
certifi = "^2024.8.30"
botocore = "^1.35.39"
requests = "^2.32.3"
wrapt = "^1.16.0"
hvac = "^2.3.0"
faker = "^30.8.1"
isort = "6.0.0"
chardet = "5.2.0"


[tool.poetry.group.dev.dependencies]
coverage = "^7.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
